+++ test/purity.jl
@@ -5,7 +5,7 @@ using Logging
 @testset "Purity checks" begin
     @testset "Qubits (d=2)" begin
         # Pure product state |00⟩ (commuting, independent, m == n).
-        tab_pure = zeros(Int64, 5, 2)
+        tab_pure = zeros(Int, 5, 2)
         tab_pure[3, 1] = 1  # Z₁
         tab_pure[4, 2] = 1  # Z₂
         # Note: is_independent is destructive (it Gauss-Jordan reduces a view),
@@ -20,7 +20,7 @@ using Logging
         @test QuditClifford.is_pure(stab_pure)
 
         # Non-commuting generators (X₁ and Z₁) should fail purity.
-        tab_noncomm = zeros(Int64, 5, 2)
+        tab_noncomm = zeros(Int, 5, 2)
         tab_noncomm[1, 1] = 1  # X₁
         tab_noncomm[3, 2] = 1  # Z₁
         stab_noncomm_comm = StabilizerTableau(2, 2, tab_noncomm; m=2, storephase=true)
@@ -34,7 +34,7 @@ using Logging
         end
 
         # Linearly dependent generators (Z₁, Z₁) should fail independence and purity.
-        tab_dep = zeros(Int64, 5, 2)
+        tab_dep = zeros(Int, 5, 2)
         tab_dep[3, 1] = 1  # Z₁
         tab_dep[3, 2] = 1  # Z₁ again
         stab_dep_comm = StabilizerTableau(2, 2, tab_dep; m=2, storephase=true)
@@ -49,7 +49,7 @@ using Logging
 
     @testset "Qudits (d=3)" begin
         # Pure product state |00⟩ (commuting, independent, m == n).
-        tab_pure = zeros(Int64, 5, 2)
+        tab_pure = zeros(Int, 5, 2)
         tab_pure[3, 1] = 1  # Z₁
         tab_pure[4, 2] = 1  # Z₂
         # Note: is_independent is destructive (it Gauss-Jordan reduces a view),
@@ -64,7 +64,7 @@ using Logging
         @test QuditClifford.is_pure(stab_pure)
 
         # Non-commuting generators (X₁ and Z₁) should fail purity.
-        tab_noncomm = zeros(Int64, 5, 2)
+        tab_noncomm = zeros(Int, 5, 2)
         tab_noncomm[1, 1] = 1  # X₁
         tab_noncomm[3, 2] = 1  # Z₁
         stab_noncomm_comm = StabilizerTableau(3, 2, tab_noncomm; m=2, storephase=true)
@@ -78,7 +78,7 @@ using Logging
         end
 
         # Linearly dependent generators (Z₁, Z₁) should fail independence and purity.
-        tab_dep = zeros(Int64, 5, 2)
+        tab_dep = zeros(Int, 5, 2)
         tab_dep[3, 1] = 1  # Z₁
         tab_dep[3, 2] = 1  # Z₁ again
         stab_dep_comm = StabilizerTableau(3, 2, tab_dep; m=2, storephase=true)
